+++ src/models/session.rs
@@ -18,7 +18,7 @@ pub struct SessionConfig {
     pub cleanup_interval_secs: u64,
     /// Maximum number of concurrent sessions per user
     #[builder(default = "5")]
-    pub max_concurrent_sessions: u32,
+    pub max_concurrent_sessions: u8,
     /// Whether session management is disabled
     #[builder(default = "false")]
     pub disable: bool,
